+++ src/cuda_oddeven_linalg_generic.h
@@ -6,28 +6,28 @@
   // 6 threads, optimized
 
   __global__ void cuda_block_oe_vector_PRECISION_copy_6threads_opt( cu_cmplx_PRECISION* out, cu_cmplx_PRECISION* in, schwarz_PRECISION_struct_on_gpu *s, 
-                                                                  int thread_id, int csw, int nr_threads_per_DD_block, int* DD_blocks_to_compute, int num_latt_site_var,
+                                                                  int thread_id, double csw, int nr_threads_per_DD_block, int* DD_blocks_to_compute, int num_latt_site_var,
                                                                   block_struct* block, int sites_to_copy );
 
-  __global__ void cuda_block_oe_vector_PRECISION_define_6threads_opt( cu_cmplx_PRECISION* spinor, schwarz_PRECISION_struct_on_gpu *s, int thread_id, int csw,
+  __global__ void cuda_block_oe_vector_PRECISION_define_6threads_opt( cu_cmplx_PRECISION* spinor, schwarz_PRECISION_struct_on_gpu *s, int thread_id, double csw,
                                                                       int nr_threads_per_DD_block, int* DD_blocks_to_compute, int num_latt_site_var, block_struct* block,
                                                                       int sites_to_define, cu_cmplx_PRECISION val_to_assign );
 
   __global__ void cuda_block_oe_vector_PRECISION_plus_6threads_opt( cu_cmplx_PRECISION* out, cu_cmplx_PRECISION* in1, cu_cmplx_PRECISION* in2,
-                                                                    schwarz_PRECISION_struct_on_gpu *s, int thread_id, int csw, int nr_threads_per_DD_block,
+                                                                    schwarz_PRECISION_struct_on_gpu *s, int thread_id, double csw, int nr_threads_per_DD_block,
                                                                     int* DD_blocks_to_compute, int num_latt_site_var, block_struct* block, int sites_to_add );
 
   __global__ void cuda_block_oe_vector_PRECISION_minus_6threads_opt( cu_cmplx_PRECISION* out, cu_cmplx_PRECISION* in1, cu_cmplx_PRECISION* in2, schwarz_PRECISION_struct_on_gpu *s,
-                                                                     int thread_id, int csw, int nr_threads_per_DD_block, int* DD_blocks_to_compute,
+                                                                     int thread_id, double csw, int nr_threads_per_DD_block, int* DD_blocks_to_compute,
                                                                      int num_latt_site_var, block_struct* block, int sites_to_add );
 
   __global__ void cuda_local_xy_over_xx_PRECISION( cu_cmplx_PRECISION* vec1, cu_cmplx_PRECISION* vec2, \
                                                  schwarz_PRECISION_struct_on_gpu *s, int thread_id, \
-                                                 int csw, int nr_threads_per_DD_block, int* DD_blocks_to_compute, \
+                                                 double csw, int nr_threads_per_DD_block, int* DD_blocks_to_compute, \
                                                  int num_latt_site_var, block_struct* block, int sites_to_dot );
 
   __global__ void cuda_block_oe_vector_PRECISION_saxpy_6threads_opt_onchip( cu_cmplx_PRECISION* out, cu_cmplx_PRECISION* in1, cu_cmplx_PRECISION* in2, PRECISION prefctr_alpha,
-                                                                            cu_cmplx_PRECISION *alpha, schwarz_PRECISION_struct_on_gpu *s, int thread_id, int csw,
+                                                                            cu_cmplx_PRECISION *alpha, schwarz_PRECISION_struct_on_gpu *s, int thread_id, double csw,
                                                                             int nr_threads_per_DD_block, int* DD_blocks_to_compute, int num_latt_site_var, block_struct* block, int sites_to_add );
